MINOR: Use composition for better isolation of fetcher logic
